requirements.txt
asgiref==3.7.2
attrs==23.1.0
awsebcli==3.20.7
bcrypt==4.0.1
blessed==1.20.0
botocore==1.29.158
cached-property==1.5.2
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.3
cryptography==41.0.2
distlib==0.3.6
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
filelock==3.12.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
jsonschema==3.2.0
paramiko==3.2.0
pathspec==0.10.1
platformdirs==3.8.1
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
requests==2.31.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
texttable==1.6.7
urllib3==1.26.16
virtualenv==20.23.1
wcwidth==0.1.9
websocket-client==0.59.0
